requirements.txt
streamlit==1.28.0
stravalib==1.6
supabase==2.0.0
pandas==2.1.0
plotly==5.17.0
python-dotenv==1.0.0